+++ vm/devices/storage/disk_nvme/nvme_driver/src/tests.rs
@@ -5,32 +5,32 @@ use crate::NvmeDriver;
 use chipset_device::mmio::ExternallyManagedMmioIntercepts;
 use chipset_device::mmio::MmioIntercept;
 use chipset_device::pci::PciConfigSpace;
-use devices_test_helpers::emulated_test_infra::EmulatedDevice;
+use devices_test_helpers::EmulatedDevice;
+use devices_test_helpers::Mapping;
 use devices_test_helpers::emulated_test_infra::GuestMemoryAccessWrapper;
-use devices_test_helpers::emulated_test_infra::Mapping;
 use guestmem::GuestMemory;
 use guid::Guid;
 use inspect::Inspect;
 use inspect::InspectMut;
 use memory_range::MemoryRange;
 use nvme::NvmeControllerCaps;
-use nvme_spec::Cap;
 use nvme_spec::nvm::DsmRange;
+use nvme_spec::Cap;
 use page_pool_alloc::PagePool;
 use page_pool_alloc::PagePoolAllocator;
 use page_pool_alloc::TestMapper;
-use pal_async::DefaultDriver;
 use pal_async::async_test;
+use pal_async::DefaultDriver;
 use parking_lot::Mutex;
 use pci_core::msi::MsiInterruptSet;
 use scsi_buffers::OwnedRequestBuffers;
 use std::sync::Arc;
 use test_with_tracing::test;
+use user_driver::interrupt::DeviceInterrupt;
+use user_driver::memory::PAGE_SIZE64;
 use user_driver::DeviceBacking;
 use user_driver::DeviceRegisterIo;
 use user_driver::DmaClient;
-use user_driver::interrupt::DeviceInterrupt;
-use user_driver::memory::PAGE_SIZE64;
 use vmcore::vm_task::SingleDriverBackend;
 use vmcore::vm_task::VmTaskDriverSource;
 use zerocopy::IntoBytes;
